+++ changelog.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `GGNLinearOperator` ([PR](https://github.com/f-dangel/curvlinops/pull/277))
   - `EFLinearOperator` ([PR](https://github.com/f-dangel/curvlinops/pull/278))
   - `GGNDiagonalLinearOperator` ([PR](https://github.com/f-dangel/curvlinops/pull/279))
+  - `JacobianLinearOperator` and `TransposedJacobianLinearOperator`
+    ([PR](https://github.com/f-dangel/curvlinops/pull/280))
   - `KFACLinearOperator` and `EKFACLinearOperator` with `backend="make_fx"`
     ([PR](https://github.com/f-dangel/curvlinops/pull/271))